+++ pkg/rm/tcc/fence/constant/tcc_fence_constant.go
@@ -17,14 +17,18 @@
 
 package constant
 
+// StatusTried
 // PHASE 1: The Commit tried.
 const StatusTried int32 = 1
 
+// StatusCommitted
 // PHASE 2: The Committed.
 const StatusCommitted int32 = 2
 
+// StatusRollbacked
 // PHASE 2: The Rollbacked.
 const StatusRollbacked int32 = 3
 
+// StatusSuspended
 // Suspended status.
 const StatusSuspended int32 = 4
